Navbar first draft
Here I've made the first draft of the Navbar component. I used inline-styling to toggle the right element of the links-container, making it appear / disappear whenever the menu icon is clicked. (The menu icon appearing on tablet-sized screens and smaller.)

I've kept the media queries for mobile in the global styles folder as mixins, then I imported them to the Navbar.scss file. This is something I'll continue to do for the styling of the rest of the components.

I haven't decided on a colour scheme yet, the colors are set to variables that can be changed.