PreviousWork Component
Here I've finished the basic structure for the PreviousWork component. I have created an array of objects which have been mapped through to display each job on the screen.

I will update the data with the relevant information later, for now it's just default data to show how everything will be displayed.