PreviousWork Enlarge image onClick
Here I've written the function to enlarge an image when it's clicked. A close icon is then displayed with an onClick to remove the large image from the screen.

Everything works as intended with one image, but the same image is displayed regardless of which one is clicked. I'll rectify this is the next commit.