PreviousWork toggleBigPic
Here I've updated the toggleBigPic function to target the picture that was clicked. So now when a picture is clicked, an enlarged version of that picture is displayed on the screen.

I did this by giving each job object 2 more properties, those being a boolean of showingEnlarged and unique key of a number. I gave the component some more state of [jobs, setJobs] and initialized it with the previousWorkData. I got rid of the setBigPic state as the showingEnlarged property did the same thing.

In the toggleBigPic function I mapped through jobs from the state and compared the id properties with the index of the job that was clicked. If a match is found that the showingEnlarged property for that job is toggled.

I used a ternary operator on the large image and the close icon, so when the showingEnlarged property for that job is set to true, the image and icon are displayed to the user. When it's false the class of 'hide' is applied removing the items from the screen.