Contact component progress
In this commit I've updated the contact component. I've made the background screenshot bigger and changed the size of this section to cover the whole screen. This seperates it from the other sections of the website making for a better user experience.

I'll carry on with the styling of this section in the next commit. This will include media quiries foe mobile screens.