Updated Navbar with Smooth-scroll / Smaller images
Here I'ved added the Links section to the navbar and have implemented smooth-scroll to each nav link. Now when a link is clicked, it scrolls to that section of the website.

I've also added an onClick={toggleMenu} to each nav link, so when they are clicked on mobile screens, the nav menu closes when the page scrolls to that section.

I've also replaced the images from the Links component with compressed versions. I'll do the same when final photos are chosen for the website.