PreviousWork Titles
Here I've given all of the data objects the correct titles for each car. Along with that I've removed the desc property from each object, as Steve only wants a title for each picture with no description. I also removed the <p> and it's styling in PreviousWork.js as it's now redundent.

 I've also adjusted the margin for the picture and title, to make for better spacing.