Bug Fixes
Here I've changed the container .info-box to have a min-height of 40vh. Before this change, some of the text went outside of it's container on small screens.

I've also changed the display for the big picture in the PreviousWork component. I've moved the x icon and big pic into a container (big-pic-container) and have centered the container to the screen. This solved the problem of the x icon and big pic moving on different screen sizes. Now everything is centered on every screen size, with the x icon in the top right corner of the big picture.

I've also given various elements different z indexes, as some elements were overlapping in the wrong places. This has now been fixed.